+++ CHANGES.rst
@@ -11,6 +11,7 @@
 - s.replace() is type annoted and documented.
 - s.replace() officially supports substitutions using a callback.
 - s.replace() code has been moved to a standalone function in utils.
+- g() check that all arguments are iterables and give a proper warning if not.
 
 
 0.1.0